Other changes. Mainly int->size_t

+++ game/src/BandwidthTest.h
@@ -32,7 +32,7 @@ public:
     void onPacketReceived(Packet* packet) {
         if(packet->getType() == PacketType::REPLY_LARGE_PACKET) {
             float delay = static_cast<float>(phantom::Util::getTime() - _requestTime);
-            int size = packet->getPayload().length();
+            size_t size = packet->getPayload().length();
 
             stringstream ss;
